+++ examples/plot_vaughan.py
@@ -27,9 +27,7 @@ from eemilib.plotter import PandasPlotter
 # Load data
 # ---------
 data_matrix = DataMatrix()
-data_matrix.set_files(
-    filepath, population="all", emission_data_type="Emission Yield"
-)
+data_matrix.set_files(filepath, population="all", data_type="Emission Yield")
 data_matrix.load_data(PandasLoader())
 plotter = PandasPlotter()
 
@@ -37,14 +35,14 @@ plotter = PandasPlotter()
 # Basic fitting
 # -------------
 axes_1 = data_matrix.plot(
-    plotter, population="all", emission_data_type="Emission Yield"
+    plotter, population="all", data_type="Emission Yield"
 )
 model_classic = Vaughan()
 model_classic.find_optimal_parameters(data_matrix)
 model_classic.plot(
     plotter,
     population="all",
-    emission_data_type="Emission Yield",
+    data_type="Emission Yield",
     energies=np.linspace(0, 1000, 1001),
     angles=np.linspace(0, 60, 4),
     axes=axes_1,
@@ -75,7 +73,7 @@ for model in models[1:]:
 # %%
 # Plot only at normal incidence to keep the plot readable.
 axes_2 = data_matrix.plot(
-    plotter, population="all", emission_data_type="Emission Yield"
+    plotter, population="all", data_type="Emission Yield"
 )
 colors = ("black", "grey", "cyan")
 linestyles = ("-", ":", "--")
@@ -83,7 +81,7 @@ for model, color, ls in zip(models, colors, linestyles):
     model.plot(
         plotter,
         population="all",
-        emission_data_type="Emission Yield",
+        data_type="Emission Yield",
         energies=np.linspace(0, 1000, 1001),
         angles=np.linspace(0, 0, 1),
         axes=axes_2,