+++ examples/test_furman_pivi.py
@@ -29,21 +29,16 @@ plotter = PandasPlotter()
 model = FurmanPivi()
 
 data_matrix.set_files(
-    teey_ag_1_heated_pd, population="all", emission_data_type="Emission Yield"
+    teey_ag_1_heated_pd, population="all", data_type="Emission Yield"
 )
-data_matrix.set_files(
-    distribs, population="all", emission_data_type="Emission Energy"
-)
+data_matrix.set_files(distribs, population="all", data_type="Emission Energy")
 data_matrix.load_data(loader)
 
 axes_distrib = data_matrix.plot(
-    plotter,
-    population="all",
-    emission_data_type="Emission Energy",
-    group_by_pe=True,
+    plotter, population="all", data_type="Emission Energy", group_by_pe=True
 )
 axes_yield = data_matrix.plot(
-    plotter, population="all", emission_data_type="Emission Yield"
+    plotter, population="all", data_type="Emission Yield"
 )
 
 model.find_optimal_parameters(data_matrix)
@@ -51,7 +46,7 @@ model.find_optimal_parameters(data_matrix)
 axes_distrib = model.plot(
     plotter,
     population=["SE", "EBE", "IBE", "all"],
-    emission_data_type="Emission Energy",
+    data_type="Emission Energy",
     energies=None,
     angles=np.linspace(0, 0, 1),
     axes=axes_distrib,
@@ -61,7 +56,7 @@ axes_distrib = model.plot(
 axes_yield = model.plot(
     plotter,
     population=["SE", "EBE", "IBE", "all"],
-    emission_data_type="Emission Yield",
+    data_type="Emission Yield",
     energies=None,
     angles=np.linspace(0, 60, 4),
     axes=axes_yield,