+++ src/eemilib/model/dionne.py
@@ -61,7 +61,7 @@ class DionneParameters(TypedDict):
 class Dionne(Model):
     """Define the Dionne model :cite:`Lye1957,Dionne1973,Dionne1975`."""
 
-    emission_data_types = ["Emission Yield"]
+    data_types = ["Emission Yield"]
     populations = ["SE"]
     considers_energy = True
     is_3d = False
@@ -154,7 +154,7 @@ class Dionne(Model):
     def get_data(
         self,
         population: ImplementedPop,
-        emission_data_type: ImplementedEmissionData,
+        data_type: ImplementedEmissionData,
         energy: NDArray[np.float64],
         theta: NDArray[np.float64],
         *args,
@@ -165,10 +165,10 @@ class Dionne(Model):
         Will return a dataframe only if the |TEEY| is asked.
 
         """
-        if population != "SE" or emission_data_type != "Emission Yield":
+        if population != "SE" or data_type != "Emission Yield":
             return super().get_data(
                 population=population,
-                emission_data_type=emission_data_type,
+                data_type=data_type,
                 energy=energy,
                 theta=theta,
                 *args,