+++ src/eemilib/model/vaughan.py
@@ -56,7 +56,7 @@ class VaughanParameters(TypedDict):
 class Vaughan(Model):
     """Define the classic Vaughan model."""
 
-    emission_data_types = ["Emission Yield"]
+    data_types = ["Emission Yield"]
     populations = ["all"]
     considers_energy = True
     is_3d = True
@@ -250,7 +250,7 @@ class Vaughan(Model):
     def get_data(
         self,
         population: ImplementedPop,
-        emission_data_type: ImplementedEmissionData,
+        data_type: ImplementedEmissionData,
         energy: NDArray[np.float64],
         theta: NDArray[np.float64],
         *args,
@@ -264,10 +264,10 @@ class Vaughan(Model):
             This method could be so much simpler and efficient.
 
         """
-        if population != "all" or emission_data_type != "Emission Yield":
+        if population != "all" or data_type != "Emission Yield":
             return super().get_data(
                 population=population,
-                emission_data_type=emission_data_type,
+                data_type=data_type,
                 energy=energy,
                 theta=theta,
                 *args,